test: align runtime tests with composition lane
